Use openFeign to achieve interact CRUD among user, post, comment and, category

+++ blogpost_service/src/main/java/com/evan/blogpost_service/service/impl/PostServiceImpl.java
@@ -1,62 +1,83 @@
 package com.evan.blogpost_service.service.impl;
 
 import com.evan.blogpost_service.dto.APIResponseDTO;
+import com.evan.blogpost_service.dto.CategoryDTO;
 import com.evan.blogpost_service.dto.PostDTO;
 import com.evan.blogpost_service.dto.UserDTO;
 import com.evan.blogpost_service.entity.Post;
-import com.evan.blogpost_service.exception.ResourceDuplicateException;
 import com.evan.blogpost_service.exception.ResourceNotFoundException;
 import com.evan.blogpost_service.mapper.PostMapper;
 import com.evan.blogpost_service.repository.PostRepository;
-import com.evan.blogpost_service.service.ApiClient;
+import com.evan.blogpost_service.service.CategoryClient;
+import com.evan.blogpost_service.service.UserClient;
 import com.evan.blogpost_service.service.PostService;
 import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.reactive.function.client.WebClient;
 
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 
 @Service
 @AllArgsConstructor
 public class PostServiceImpl implements PostService {
     private static final Logger LOGGER = LoggerFactory.getLogger(PostServiceImpl.class);
-//    private WebClient webClient;
-    private ApiClient apiClient;
+    private UserClient userClient;
+    private CategoryClient categoryClient;
     private PostRepository postRepository;
 
     @Override
-    public PostDTO savePost(PostDTO post) {
+    public APIResponseDTO savePost(PostDTO post) {
         LOGGER.info("in savePost()");
+        // to get current user email
+        String author = "user1@gmail.com";
+        post.setAuthor(author);
+        UserDTO userDTO = userClient.getUser(author);
+        if (userDTO == null) throw new ResourceNotFoundException("Author", "email", author);
+
         Post convertPost = PostMapper.mapToPost(post);
         Post newPost = postRepository.save(convertPost);
         PostDTO postDto = PostMapper.mapToPostDTO(newPost);
-        return postDto;
+
+        // TODO: append postId into current user
+        Set<Long> addPost = userDTO.getPosts();
+        addPost.add(postDto.getPostId());
+        userDTO.setPosts(addPost);
+        LOGGER.info("after add post in userDTO: " + userDTO.getPosts() + " the user email is: " + userDTO.getEmail());
+        userClient.updateUser(userDTO);
+        UserDTO newUserDTO = userClient.getUser(author);
+        LOGGER.info("the newest user: " + newUserDTO.getPosts() + " the email is: : " + newUserDTO.getEmail());
+
+//        Append category which current post belong into category database
+        CategoryDTO categoryDTO = categoryClient.getCategoryByName(newPost.getCategory());
+        if (categoryDTO == null) throw new ResourceNotFoundException("Category", "name", newPost.getCategory());
+        Set<Long> addPostToCategory = categoryDTO.getPostIds();
+        addPostToCategory.add(newPost.getPostId());
+        categoryDTO.setPostIds(addPostToCategory);
+        categoryClient.updateCategory(categoryDTO);
+        CategoryDTO newCategoryDTO = categoryClient.getCategoryByName(newPost.getCategory());
+        LOGGER.info("add current postId into category and update success: " + newCategoryDTO.getPostIds());
+
+        APIResponseDTO apiResponseDTO = new APIResponseDTO();
+        apiResponseDTO.setUserDTO(newUserDTO);
+        apiResponseDTO.setPostDTO(postDto);
+        apiResponseDTO.setCategoryDTO(newCategoryDTO);
+        LOGGER.info("success store response into apiResponse, the userDTO is: " + apiResponseDTO.getUserDTO().getPosts() + " the categoryDTO is: " + apiResponseDTO.getCategoryDTO().getPostIds());
+        return apiResponseDTO;
     }
 
     @Override
-    public APIResponseDTO getPostById(Long postId) {
-        LOGGER.info("in getPostById");
-        Post post = postRepository.findById(postId).get();
+    public PostDTO getPostById(Long postId) {
         if (postId == null) throw new ResourceNotFoundException("Post", "id", postId);
+        LOGGER.info("in getPostById, the postId is: " + postId);
+        Post post = postRepository.findById(postId).get();
+        if (post == null) throw new ResourceNotFoundException("Post", "id", postId);
 
-//        UserDTO userDTO = webClient.get()
-//                .uri("http://localhost:8081/api/users/get/" + post.getAuthor())
-//                .retrieve()
-//                .bodyToMono(UserDTO.class)
-//                .block();
-        UserDTO userDTO = apiClient.getUser(post.getAuthor());
         PostDTO postDto = PostMapper.mapToPostDTO(post);
 
-        APIResponseDTO apiResponseDTO = new APIResponseDTO();
-        apiResponseDTO.setUserDTO(userDTO);
-        apiResponseDTO.setPostDTO(postDto);
-
-//        return postDto;
-        return apiResponseDTO;
+        return postDto;
     }
 
     @Override
@@ -85,7 +106,7 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public PostDTO updatePost(PostDTO postDTO) {
-        LOGGER.info("in updatePost()");
+        LOGGER.info("in updatePost(), the postId is: " + postDTO.getPostId());
         Post oldPost = postRepository.findById(postDTO.getPostId()).get();
         if (oldPost == null) throw new ResourceNotFoundException("Post", "id", postDTO.getPostId());
         Post newPost = PostMapper.mapToPost(postDTO);
@@ -100,12 +121,48 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
-    public String deletePost(Long postId) {
+    public APIResponseDTO deletePost(Long postId) {
         LOGGER.info("in deletePost()");
         Post delPost = postRepository.findById(postId).get();
+        PostDTO delPostDTO = PostMapper.mapToPostDTO(delPost);
+        LOGGER.info("the post which will be deleted is: " + delPostDTO.getPostId(), " the correspond category name is: " + delPostDTO.getCategory());
+
+//        check the post if is existed:
         if (delPost == null) throw new ResourceNotFoundException("Post", "id", postId);
+
+//        get author information, so that we can delete postId from the author
+        String author = delPostDTO.getAuthor();
+        UserDTO userDTO = userClient.getUser(author);
+
+        if (userDTO == null) throw new ResourceNotFoundException("Author", "email", author);
         postRepository.delete(delPost);
-        return "Success delete";
+
+        // TODO: delete postId which from user
+//        delete post from user then update
+        Set<Long> delPostFromUser = userDTO.getPosts();
+        if (delPostFromUser.contains(postId)) delPostFromUser.remove(postId);
+        userDTO.setPosts(delPostFromUser);
+//        get new user information
+        UserDTO newUserDTO = userClient.updateUser(userDTO);
+        LOGGER.info("success update user data: " + newUserDTO.getPosts());
+
+//        delete postId from category
+        CategoryDTO categoryDTO = categoryClient.getCategoryByName(delPostDTO.getCategory());
+        if (categoryDTO == null) throw new ResourceNotFoundException("Category", "name", delPostDTO.getCategory());
+        Set<Long> delPostFromCategory = categoryDTO.getPostIds();
+        LOGGER.info("the original category's postIds is: " + delPostFromCategory);
+
+        if (delPostFromCategory.contains(delPostDTO.getPostId())) delPostFromCategory.remove(postId);
+        categoryDTO.setPostIds(delPostFromCategory);
+        CategoryDTO newCategoryDTO = categoryClient.updateCategory(categoryDTO);
+        LOGGER.info("after delete the correspond postId, the category's postIds is: " + newCategoryDTO.getPostIds());
+
+        APIResponseDTO apiResponseDTO = new APIResponseDTO();
+        apiResponseDTO.setUserDTO(newUserDTO);
+        apiResponseDTO.setCategoryDTO(newCategoryDTO);
+        LOGGER.info("Post deleted successfully: newUser: " + newUserDTO.getPosts() + " newCategoryDTO: " + newCategoryDTO.getPostIds());
+
+        return apiResponseDTO;
     }
 
 }
